# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Condominium",
        "description": "API documentation for the Condominium management system",
        "version": "1.0"
    },
    "paths": {
        "/api/owners": {
            "get": {
                "tags": [
                    "Owners"
                ],
                "summary": "Listar todos os moradores",
                "description": "Retorna uma lista de moradores paginados",
                "operationId": "9e6af3de1e756f8f19aa17b9f190d639",
                "responses": {
                    "200": {
                        "description": "Lista de moradores retornada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "string",
                                                        "example": "1"
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "João Silva"
                                                    },
                                                    "cpf": {
                                                        "type": "string",
                                                        "example": "12345678900"
                                                    },
                                                    "birth_date": {
                                                        "type": "string",
                                                        "format": "date",
                                                        "example": "1990-01-01"
                                                    },
                                                    "email": {
                                                        "type": "string",
                                                        "example": "joao@exemplo.com"
                                                    },
                                                    "tower": {
                                                        "type": "string",
                                                        "example": "A"
                                                    },
                                                    "apartment_number": {
                                                        "type": "string",
                                                        "example": "101"
                                                    },
                                                    "garage": {
                                                        "type": "string",
                                                        "example": "1"
                                                    },
                                                    "rented": {
                                                        "type": "boolean",
                                                        "example": true
                                                    },
                                                    "paid": {
                                                        "type": "boolean",
                                                        "example": true
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Lista de moradores retornada com sucesso"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autorizado"
                    }
                }
            },
            "post": {
                "tags": [
                    "Owners"
                ],
                "summary": "Cadastrar um morador",
                "description": "Cadastra um morador no sistema",
                "operationId": "f6d6bf3981a1f44f42bb408ade192f8e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "cpf",
                                    "birth_date",
                                    "email",
                                    "tower",
                                    "apartment_number",
                                    "garage",
                                    "rented",
                                    "paid"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "João Silva"
                                    },
                                    "cpf": {
                                        "type": "string",
                                        "example": "12345678900"
                                    },
                                    "birth_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "1990-01-01"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "joao@exemplo.com"
                                    },
                                    "tower": {
                                        "type": "string",
                                        "example": "A"
                                    },
                                    "apartment_number": {
                                        "type": "string",
                                        "example": "101"
                                    },
                                    "garage": {
                                        "type": "string",
                                        "example": "1"
                                    },
                                    "rented": {
                                        "type": "boolean",
                                        "example": true
                                    },
                                    "paid": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Morador cadastrado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Morador cadastrado com sucesso!"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "string",
                                                    "example": "1"
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "João Silva"
                                                },
                                                "cpf": {
                                                    "type": "string",
                                                    "example": "12345678900"
                                                },
                                                "birth_date": {
                                                    "type": "string",
                                                    "format": "date",
                                                    "example": "1990-01-01"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "joao@exemplo.com"
                                                },
                                                "tower": {
                                                    "type": "string",
                                                    "example": "A"
                                                },
                                                "apartment_number": {
                                                    "type": "string",
                                                    "example": "101"
                                                },
                                                "garage": {
                                                    "type": "string",
                                                    "example": "1"
                                                },
                                                "rented": {
                                                    "type": "boolean",
                                                    "example": true
                                                },
                                                "paid": {
                                                    "type": "boolean",
                                                    "example": true
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "CPF inválido"
                    },
                    "500": {
                        "description": "Erro ao cadastrar morador"
                    }
                }
            }
        },
...
    },
...
}